Kodus veits 20-10-01

+++ src/main/java/com/VeebiPood/service/ShopRepository.java
@@ -82,17 +82,30 @@ public class ShopRepository {
         jdbcTemplate.update(sql, paramMap);
     }
 
-    // TODO
-    public void removeItemFromCart(Long productId, Long accountId, Long quantity, BigDecimal price) {
-        String sql = "DELETE FROM cart_item WHERE product_id = :product_id, ;";
+    public void updateQuantity(Long productId, Long accountId, Long newQuantity) {
+        String sql = "UPDATE cart_item SET quantity = :quantity WHERE account_id = :account_id, ";
         Map<String, Object> paramMap = new HashMap<>();
         paramMap.put("product_id", productId);
         paramMap.put("account_id", accountId);
-        paramMap.put("quantity", quantity);
-        paramMap.put("price", price);
+        paramMap.put("quantity", newQuantity);
         jdbcTemplate.update(sql, paramMap);
     }
 
+    public void removeItemFromCart(Long productId, Long accountId) {
+        String sql = "DELETE FROM cart_item WHERE account_id = :account_id AND product_id = :product_id";
+        Map<String, Object> paramMap = new HashMap<>();
+        paramMap.put("product_id", productId);
+        paramMap.put("account_id", accountId);
+        jdbcTemplate.update(sql, paramMap);
+    }
+
+    public Long checkIfItemInCart(Long productId, Long accountId) {
+        String sql = "SELECT quantity FROM cart_item WHERE product_id = :product_id, account_id = :account_id";
+        Map<String, Object> paramMap = new HashMap<>();
+        paramMap.put("product_id", productId);
+        paramMap.put("account_id", accountId);
+        return jdbcTemplate.queryForObject(sql, new HashMap<>(), Long.class);
+    }
 
     public Long getAccountId(String userName) {
         String sql = "SELECT id FROM account WHERE user_name = :user_name";
@@ -107,6 +120,8 @@ public class ShopRepository {
         paramMap.put("account_id", accountId);
         return jdbcTemplate.query(sql, paramMap, new CartItemListRowMapper());
     }
+
+
 
 
 //    public List<CartItem> getCartItemList(CartItem cartItem) {
